// src/services/types/platform/follow_up_close.ts
export type FollowUpCloseApiResponse = {
  _id: string
  __v: number
  follow_up_and_close: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
    }
    section3: {
      image1: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
      text4_en: string
      text4_fa: string
      text5_en: string
      text5_fa: string
      image1: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      image1: string
    }
    section6: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
    }
    section7: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}